+++ src/lib.rs
@@ -1,4 +1,2 @@
-/// Main directory containing all db stuff
 pub mod db;
-
 pub use bevy_reflect::Reflect;